refactor(Calendar): Move Combobox comp into base components and split component into separate files
